requirements.txt
rti.connext==7.6.0
PySide6==6.7.*
pytest==8.2.*
pytest-asyncio>=1.0
pytest-qt==4.4.*
pytest-xdist[psutil]==3.5.*
black==24.*
isort==5.13.*
ruff==0.5.*
